working on test cases

+++ support/cas-server-support-couchdb-ticket-registry/src/main/java/org/apereo/cas/couchdb/tickets/TicketRepository.java
@@ -31,8 +31,10 @@ public class TicketRepository extends CouchDbRepositorySupport<TicketDocument> {
      * @return number of tickets deleted
      */
     public long delete(final List<TicketDocument> ticketDocuments) {
-        return (long) db.executeBulk(ticketDocuments.stream().map(BulkDeleteDocument::of)
-            .collect(Collectors.toList())).size();
+        return db.executeBulk(ticketDocuments.stream()
+            .map(BulkDeleteDocument::of)
+            .collect(Collectors.toList()))
+            .size();
     }
 
     /**
